+++ src/main/kotlin/com/wutsi/codegen/kotlin/server/ServerGithubActionsCodeGenerator.kt
@@ -1,17 +1,14 @@
 package com.wutsi.codegen.kotlin.server
 
 import com.wutsi.codegen.Context
-import com.wutsi.codegen.core.generator.AbstractMustacheCodeGenerator
 import com.wutsi.codegen.core.util.CaseUtil
+import com.wutsi.codegen.github.AbstractGithubActionsCodeGenerator
 import com.wutsi.codegen.kotlin.KotlinMapper
 import io.swagger.v3.oas.models.OpenAPI
-import java.io.File
 
-class ServerGithubActionsCodeGenerator : AbstractMustacheCodeGenerator() {
-    override fun generate(openAPI: OpenAPI, context: Context) {
-        generate("master.yml", openAPI, context)
-        generate("pull_request.yml", openAPI, context)
-    }
+class ServerGithubActionsCodeGenerator : AbstractGithubActionsCodeGenerator() {
+    override fun getInputFilePath(filename: String): String =
+        "/kotlin/server/.github/workflows/$filename.mustache"
 
     override fun toMustacheScope(openAPI: OpenAPI, context: Context) = mapOf(
         "artifactId" to ServerMavenCodeGenerator(KotlinMapper(context)).artifactId(context),
@@ -51,10 +48,5 @@ class ServerGithubActionsCodeGenerator : AbstractMustacheCodeGenerator() {
             result["cache"] = true
         }
         return result
-    }
-
-    private fun generate(filename: String, openAPI: OpenAPI, context: Context) {
-        val file = File(context.outputDirectory + File.separator + ".github" + File.separator + "workflows" + File.separator + filename)
-        generate("/.github/workflows/$filename.mustache", file, openAPI, context)
     }
 }